- we handle the value given in the request body as a string because of validation purpose (see details in service methods) - adds logging functionality - uses BigInteger for fibonacci value computation. - Adjusts tests

+++ src/main/java/com/example/cimt/controller/MathController.java
@@ -7,29 +7,28 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
+
 @RestController
-@RequestMapping ("/values")
+@RequestMapping("/values")
 public class MathController {
 
-	@Autowired
-	MathService mathService;
+    @Autowired
+    MathService mathService;
 
 
-	@GetMapping("/fibonacci/{position}")
-	public ResponseEntity<Integer> getFibonacciNumber(@PathVariable int position) {
-		 int fiboValue = mathService.calculateFiboValue(position);
-		 return ResponseEntity.ok(fiboValue);
-	}
+    @GetMapping("/fibonacci/{position}")
+    public ResponseEntity<BigInteger> getFibonacciNumber(@PathVariable int position) {
+        BigInteger fiboValue = mathService.calculateFiboValue(position);
+        return ResponseEntity.ok(fiboValue);
+    }
 
 
-
-	@PostMapping("/gcd")
-	public ResponseEntity<Integer> calculateGCD(@RequestBody @Valid GcdRequestBody request) {
-
-		int gcdValue = mathService.calculateGcd(request.getVal1(), request.getVal2());
-		return ResponseEntity.ok(gcdValue);
-	}
-
+    @PostMapping("/gcd")
+    public ResponseEntity<Integer> calculateGCD(@RequestBody @Valid GcdRequestBody request) {
+        int gcdValue = mathService.calculateGcd(request.getVal1(), request.getVal2());
+        return ResponseEntity.ok(gcdValue);
+    }
 
 
 }
